requirements.txt
attrs==22.2.0
click==8.1.3
colorama==0.4.6
coverage==7.2.2
cycler==0.11.0
exceptiongroup==1.1.1
Flask==2.2.2
fonttools==4.38.0
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.5.3
numpy==1.24.2
packaging==23.0
pandas==1.1.5
Pillow==9.4.0
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.2
pytest-cov==4.0.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
scikit-surprise==1.1.1
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
tomli==2.0.1
typing-extensions==4.4.0
Werkzeug==2.2.2
zipp==3.13.0